Criado middleware para controle de acesso

+++ routes/web.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Core\Middleware\RedirectWithoutPermission;
 use App\Http\CollaboratorController;
 use App\Http\HomeController;
 use Illuminate\Support\Facades\Auth;
@@ -20,7 +21,7 @@ Route::group(['prefix' => 'home', 'middleware' => 'auth'], function () {
         Route::post('/store', [CollaboratorController::class, 'store'])->name('home.collaborators.store');
         Route::get('/edit/{id}', [CollaboratorController::class, 'edit'])->name('home.collaborators.edit');
         Route::put('/update/{id}', [CollaboratorController::class, 'update'])->name('home.collaborators.update');
-        Route::delete('/delete/{id}', [CollaboratorController::class, 'delete'])->name('home.collaborators.delete');
-        Route::post('/batch', [CollaboratorController::class, 'batch'])->name('home.collaborators.batch');
+        Route::delete('/delete/{id}', [CollaboratorController::class, 'delete'])->middleware(RedirectWithoutPermission::class)->name('home.collaborators.delete');
+        Route::post('/batch', [CollaboratorController::class, 'batch'])->middleware(RedirectWithoutPermission::class)->name('home.collaborators.batch');
     });
 });
